fix: Added financial year selection to new/edit transaction page
This forces a selection from the user for which financial year should be used.
Once a financial year has been selected, the transaction received date datepicker's minimal date is set to the start date of the financial year, making it impossible to add transactions in nonexisting financial years.
Closed financial years can not be selected, making it impossible to add new or change existing transactions to closed financial years.